Tests/LibWeb/Text/expected/URL/search-params-delete.txt

animal=octopus&place=ocean
animal=octopus&place=ocean
place=ocean
place=ocean
